moved EditorBuffer to mod, emoji
